Add case for already installed link

+++ install.sh
@@ -4,14 +4,19 @@ TARGET="$HOME/.nix-defexpr/local"
 SOURCE="$(pwd)"
 
 if [ -e "$TARGET" ]; then
-  echo "Target file ($TARGET) exist..."
-  echo -n "Do you wanna delete it (y/n)? "
-  read yesno
-  if [ "$yesno" = "Y" -o "$yesno" = "y" ]; then
-    rm -f "$TARGET"
-  else
-    echo "Exiting installation..."
+  if [ "$(readlink $TARGET)" = "$SOURCE" ]; then
+    echo "Already installed, exiting..."
     exit 1
+  else
+    echo "Target file ($TARGET) exist..."
+    echo -n "Do you wanna delete it (y/n)? "
+    read yesno
+    if [ "$yesno" = "Y" -o "$yesno" = "y" ]; then
+      rm -f "$TARGET"
+    else
+      echo "Exiting installation..."
+      exit 1
+    fi
   fi
 fi
